Querying Clients

This section will take you through setting a client and querying it.

In order to query data, this package must have access to a valid GraphQL client. By default, this client is the gateway (https://api.thegraph.com/subgraphs/name/graphprotocol/graph-network-mainnet). However, you can change the client to another.

julia> client!("mynewclient.com/endpoint")

TheGraphData.client! — Function

client!(u::AbstractString)
client!(c::GraphQLClient.Client)

Set the provided url (u) or client (c) as the new global client.

This sets the global client, so all subsequent calls to querying function do not need you to track and pass the client around.

After ensuring that you've correctly set the client, you can now query the client for data.

julia> qvalue = "subgraphDeployments"
julia> qargs = Dict("first" => 1000, "orderBy" => "signalledTokens")
julia> qfields = ["ipfsHash"]
julia> query(qvalue, qargs, qfields)

TheGraphData.query — Function

query(v::AbstractString, a::Dict, f::AbstractVector{S}) where {S<:AbstractString}

Query the client for value v with arguments a and fields f.

By default, the client is the gateway. This function returns a vector of dictionaries.